+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ This document describes changes between each past release.
 13.7.0 (unreleased)
 -------------------
 
-- Nothing changed yet.
+**Security fix**
+
+- Fix permission checking with POST on plural endpoints (fixes #2459)
 
 
 13.6.3 (2020-01-29)