+++ CHANGELOG.rst
@@ -10,6 +10,15 @@ This document describes changes between each past release.
 
 - Fix permission checking with POST on plural endpoints (fixes #2459)
 
+**Bug fixes**
+
+- Apply CORS headers to EOL responses (#2452)
+
+**Internal Changes**
+
+- Remove ``auth`` parameter from postgresql/memory storage backends
+- Removed pytest-sugar
+
 
 13.6.3 (2020-01-29)
 -------------------